models: add conversion support for org models

+++ Src/CSharpRefImpl.Models/convert.sh
@@ -8,9 +8,9 @@
 #
 # Ensure we have the expected number of arguments.
 #
-if [ $# != 1 ]; then
-    echo "USAGE: DOGEN_BINARY" >&2
-    exit 0;
+if [ $# != 2 ]; then
+    echo "USAGE: DOGEN_BINARY CODEC" >&2
+    exit 1;
 fi
 
 #
@@ -19,7 +19,7 @@ fi
 dogen_binary=$1;
 if [ ! -f ${dogen_binary} ]; then
     echo "Could not find dogen binary."
-    exit 0;
+    exit 1;
 fi
 
 #
@@ -41,10 +41,19 @@ dogen_general_options="${dogen_general_options} --byproduct-directory ${dogen_lo
 frontend="dia"
 models="CSharpRefImpl.CSharpModel CSharpRefImpl.DirectorySettings CSharpRefImpl.LamModel CSharpRefImpl.Profiles";
 for model in ${models}; do
-    echo "Converting ${model}";
-    dogen_source_option="--source ${script_dir}/dia/${model}.${frontend}";
-    dogen_destination_option="--destination ${script_dir}/json/${model}.tmp.json";
-    ${dogen_binary} convert ${dogen_general_options} ${dogen_source_option} ${dogen_destination_option}
-    jq . ${script_dir}/json/${model}.tmp.json > ${script_dir}/json/${model}.json
-    rm ${script_dir}/json/${model}.tmp.json
+    echo "Converting ${model} to $2";
+    if [ "$2" = "json" ]; then
+        dogen_source_option="--source ${script_dir}/dia/${model}.${frontend}";
+        dogen_destination_option="--destination ${script_dir}/json/${model}.tmp.json";
+        ${dogen_binary} convert ${dogen_general_options} ${dogen_source_option} ${dogen_destination_option}
+        jq . ${script_dir}/json/${model}.tmp.json > ${script_dir}/json/${model}.json
+        rm ${script_dir}/json/${model}.tmp.json
+    elif [ "$2" = "org" ]; then
+        dogen_source_option="--source ${script_dir}/dia/${model}.${frontend}";
+        dogen_destination_option="--destination ${script_dir}/org/${model}.org";
+        ${dogen_binary} convert ${dogen_general_options} ${dogen_source_option} ${dogen_destination_option}
+    else
+        echo "Unsupported codec: " $2
+        exit 1;
+    fi
 done
